chained database and fixed frontend view

+++ app/Http/Controllers/Admin/PostController.php
@@ -2,33 +2,41 @@
 
 namespace App\Http\Controllers\Admin;
 
-use Illuminate\Validation\Rule;
 use App\Http\Controllers\Controller;
 use App\Post;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Routing\Route;
+
 
 class PostController extends Controller
 {
-    protected $validationRules = [
-        'title',
-        'slug',
-        'description'
+    public $validators = [
+        'title'     => 'required|max:100',
+        'content'   => 'required'
     ];
 
-    public function getValidation($value)
+    private function getValidators($model)
     {
-        return
-            [
-                'title'     => 'required|max:100',
-                'slug' => [
-                    'required',
-                    Rule::unique('posts')->ignore($value),
-                    'max:100'
-                ],
-                'content'   => 'required'
-            ];
+        return [
+            // 'user_id'   => 'required|exists:App\User,id',
+            'title'     => 'required|max:100',
+            'slug'      => [
+                'required',
+                Rule::unique('posts')->ignore($model),
+                'max:100'
+            ],
+            'content'   => 'required'
+        ];
     }
 
+    public function myindex()
+    {
+        $posts = Post::where('user_id', Auth::user()->id)->paginate(50);
+
+        return view('admin.posts.index', compact('posts'));
+    }
 
     /**
      * Display a listing of the resource.
@@ -37,7 +45,6 @@ class PostController extends Controller
      */
     public function index()
     {
-        //
         $posts = Post::paginate(50);
 
         return view('admin.posts.index', compact('posts'));
@@ -50,8 +57,6 @@ class PostController extends Controller
      */
     public function create()
     {
-        //
-
         return view('admin.posts.create');
     }
 
@@ -61,12 +66,14 @@ class PostController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request, Post $post)
+    public function store(Request $request)
     {
-        //
+        $request->validate($this->getValidators(null));
 
-        $request->validate($this->getValidation(null));
-        $post = Post::create($request->all());
+        $formData = $request->all() + [
+            'user_id' => Auth::user()->id
+        ];
+        $post = Post::create($formData);
 
         return redirect()->route('admin.posts.show', $post->slug);
     }
@@ -79,8 +86,6 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        //
-
         return view('admin.posts.show', compact('post'));
     }
 
@@ -92,8 +97,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        //
-
+        if (Auth::user()->id !== $post->user_id) abort(403);
         return view('admin.posts.edit', compact('post'));
     }
 
@@ -106,9 +110,10 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        //
+        if (Auth::user()->id !== $post->user_id) abort(403);
 
-        $request->validate($this->getValidation($post));
+        $request->validate($this->getValidators($post));
+
         $post->update($request->all());
 
         return redirect()->route('admin.posts.show', $post->slug);
@@ -122,10 +127,13 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        //
+        if (Auth::user()->id !== $post->user_id) abort(403);
 
         $post->delete();
 
-        return redirect()->route('admin.posts.index');
+        if (url()->previous() === route('admin.posts.edit', $post->slug)) {
+            return redirect()->route('admin.home')->with('status', "Post $post->title deleted");;
+        }
+        return redirect(url()->previous())->with('status', "Post $post->title deleted");;
     }
 }
